style: move history and bookmarks to left
