+++ src/renderer/views/app/components/Overlay/views/bookmarks/style.ts
@@ -6,11 +6,3 @@ export const BookmarkSection = styled(Section)`
   padding: 8px 0px;
   margin-top: 48px;
 `;
-
-export const Sections = styled.div`
-  margin-left: 300px;
-  width: calc(100% - 300px);
-  display: flex;
-  flex-flow: column;
-  align-items: center;
-`;
